fastapi_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
pymysql
python-multipart
python-dotenv
